bot/requirements.txt
dotenv
discord.py
langchain
langchain-ollama